requirements.txt
streamlit==1.28.1
langchain==0.1.20
langchain-community==0.0.38
langchain-google-genai==1.0.1
google-generativeai>=0.3.0,<0.5.0
chromadb==0.4.18
pypdf==3.17.4
python-dotenv==1.0.0
tiktoken>=0.5.0,<0.8.0
sentence-transformers==2.2.2
streamlit-chat==0.1.1
pandas==2.1.4
numpy==1.24.3
